house-management-api/requirements.txt
FastAPI
uvicorn
pydantics
python-dotenv
pytest
httpx